Don't load videos if zoomVideos is false
Fixes #75. If Zoom videos in options is set to false, stop loading of
the video preview. Sometimes hoverzoom got stuck in the loading-phase,
because it started loading a video preview without knowing it was a
video, and when it discovered it was a video it didn't stop the loading
properly. Here, we stop the loading as soon as we know it's a video.

After this change, clips from clips.twitch.tv was still leaving a
loading spinner when not zooming videos, so I added a check in the
plugin that respects the options, and loads an image instead of the
video when not zooming videos.